Fix various minor bugs
Fixed the menu items, they appear in the correct order and as icons
as appropriate now.

Fixed autosave crashing bug.

Adjusted file dialog layout.

+++ app/src/main/java/com/jaween/pixelart/tools/Pen.java
@@ -24,8 +24,8 @@ public class Pen extends Tool {
 
     @Override
     protected void onStart(Bitmap bitmap, PointF event) {
-        start.x = event.x;
-        start.y = event.y;
+        start.x = (int) event.x;
+        start.y = (int) event.y;
 
         draw(bitmap, event);
     }
